test(playground): move the media run-path tests to their own file
`getChatCompletionInput.test.ts` is upstream's, and the fork had appended a describe
block to it. Its own file instead, so upstream's is byte-identical again and the two
can be edited without meeting.

That leaves `spanDownloadUtils.test.ts` as the only upstream test file the fork
touches, and only because it genuinely cannot compile otherwise: the media endpoints
widened the generic on `authApiFetch.GET`, changing which overload TypeScript picks
for an inline mock literal. What remains there is an import and three one-line swaps.
